Trim a few bits
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@8180 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libpocketsphinx/ngram_search.h
@@ -97,9 +97,9 @@ typedef struct root_chan_s {
     int32  this_phn_wid;	/**< list of words consisting of this single phone;
 				   actually the first of the list, like penult_phn_wid;
 				   -1 if none */
-    int32    diphone;		/**< first diphone of this node; all words rooted at this
+    int16    diphone;		/**< first diphone of this node; all words rooted at this
 				   node begin with this diphone */
-    int32    ciphone;		/**< first ciphone of this node; all words rooted at this
+    int16    ciphone;		/**< first ciphone of this node; all words rooted at this
 				   node begin with this ciphone */
 } root_chan_t;
